Att
Atualization of queries and removal of painel_model
